+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -566,7 +566,7 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     }
 
     /**
-     * @return the first element which the [selector] yields the smallest value or null if empty
+     * @return the first element which the [selector] yields the smallest value, or null if empty
      */
     public inline fun <R : Comparable<R>> minByOrNull(selector: (element: T) -> R): T? {
         return if (isEmpty()) null else minBy(selector)
@@ -590,6 +590,13 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
             }
         }
         return maxElement
+    }
+
+    /**
+     * @return the first element which the [selector] yields the largest value, or null if empty
+     */
+    public inline fun <R : Comparable<R>> maxByOrNull(selector: (element: T) -> R): T? {
+        return if (isEmpty()) null else maxBy(selector)
     }
 
     /**