+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableDoubleArray.kt
@@ -563,7 +563,7 @@ public value class ImmutableDoubleArray @PublishedApi internal constructor(
     }
 
     /**
-     * @return the first element which the [selector] yields the smallest value or null if empty
+     * @return the first element which the [selector] yields the smallest value, or null if empty
      */
     public inline fun <R : Comparable<R>> minByOrNull(selector: (element: Double) -> R): Double? {
         return if (isEmpty()) null else minBy(selector)
@@ -587,6 +587,13 @@ public value class ImmutableDoubleArray @PublishedApi internal constructor(
             }
         }
         return maxElement
+    }
+
+    /**
+     * @return the first element which the [selector] yields the largest value, or null if empty
+     */
+    public inline fun <R : Comparable<R>> maxByOrNull(selector: (element: Double) -> R): Double? {
+        return if (isEmpty()) null else maxBy(selector)
     }
 
     /**